update make-mesa-symlinks.sh for dril

+++ make-mesa-symlinks.sh
@@ -17,14 +17,26 @@ create_links()
         exit 1
     fi
 
-    ln -sf $mesalibs/src/glx/libGLX_mesa.so             $libdir/libGLX_mesa.so.0.0.0
-    ln -sf $mesalibs/src/egl/libEGL_mesa.so             $libdir/libEGL_mesa.so.0.0.0
-    ln -sf $mesalibs/src/mapi/shared-glapi/libglapi.so  $libdir/libglapi.so.0.0.0
-    ln -sf $mesalibs/src/gbm/libgbm.so                  $libdir/libgbm.so.1.0.0
+    cd $mesalibs/src/gallium/targets/dri
+    libgallium_so=`echo libgallium-*.so`
+    cd - >/dev/null
 
-    ln -sf $mesalibs/src/gallium/targets/dri/libgallium_dri.so    $libdir/dri/radeonsi_dri.so
-    ln -sf $mesalibs/src/gallium/targets/dri/libgallium_dri.so    $libdir/dri/swrast_dri.so
-    ln -sf $mesalibs/src/gallium/targets/dri/libgallium_dri.so    $libdir/dri/zink_dri.so
+    if test ! -f "$mesalibs/src/gallium/targets/dri/$libgallium_so"; then
+        echo Expected $mesalibs/src/gallium/targets/libgallium-*.so. Found: "$libgallium_so" '(or the file is missing)'
+        exit 1
+    fi
+
+    ln -sf $mesalibs/src/egl/libEGL_mesa.so             $libdir/libEGL_mesa.so.0.0.0
+    ln -sf $mesalibs/src/glx/libGLX_mesa.so             $libdir/libGLX_mesa.so.0.0.0
+    ln -sf $mesalibs/src/gbm/libgbm.so                  $libdir/libgbm.so.1.0.0
+    ln -sf $mesalibs/src/mapi/shared-glapi/libglapi.so  $libdir/libglapi.so.0.0.0
+
+    ln -sf $mesalibs/src/gallium/targets/dri/$libgallium_so     $libdir/$libgallium_so
+
+    ln -sf $mesalibs/src/gallium/targets/dril/libdril_dri.so    $libdir/dri/libdril_dri.so
+    ln -sf $mesalibs/src/gallium/targets/dril/libdril_dri.so    $libdir/dri/radeonsi_dri.so
+    ln -sf $mesalibs/src/gallium/targets/dril/libdril_dri.so    $libdir/dri/swrast_dri.so
+    ln -sf $mesalibs/src/gallium/targets/dril/libdril_dri.so    $libdir/dri/zink_dri.so
 
     ln -sf $mesalibs/src/amd/vulkan/libvulkan_radeon.so $libdir/libvulkan_radeon.so
